refactor(license): 实施最终UI架构，统一交互处理模式
本次提交是对授权助手UI层的一次根本性、决定性的重构。我们废弃了在 `__init__` 中手动添加组件的旧模式，全面转向使用装饰器（`@ui.button`, `@ui.select`）的声明式UI。

核心改动是确立并实施了一套处理Discord交互的“黄金设计模式”：
1.  **非模态框交互**: 在回调函数入口处使用 `safe_defer()` 进行统一、安全的响应。
2.  **模态框开启**: 直接调用 `interaction.response.send_modal()` 作为唯一的初始响应。
3.  **模态框提交 (`on_submit`)**: 将其视为一个新的交互入口，同样使用 `safe_defer()`。

为支持此模式，我们引入了以下关键实践：
- **`custom_id` 全面应用**: 为所有UI组件分配了唯一的、命名空间化的`custom_id`。
- **可靠的组件查找**: 废弃了脆弱的 `self.children[n]` 索引访问，转而使用 `get_item_by_id()` 辅助函数进行稳定查找。
- **工具函数抽象**: 创建了 `safe_defer()` 和 `do_simple_owner_id_interaction_check()`，将重复的、关键的逻辑抽象出来，保证了代码的DRY（Don't Repeat Yourself）原则和稳定性。

这次重构的成果是：
- 彻底根除了所有已知的、由交互响应不当引起的 `10062: Unknown interaction` 和 `404: Not Found` 错误。
- 大幅提升了代码的可读性和可维护性，UI布局与逻辑完全分离。
- 确立了一套可被项目内所有未来UI组件复用的、稳定可靠的架构规范。
