+++ src/license/view/view_setting.py
@@ -7,8 +7,8 @@ if TYPE_CHECKING:
     from src.license.cog import LicenseCog
 from src.license.constants import ACTIVE_COMMAND_CONFIG
 from src.license.database import LicenseDB, LicenseConfig, get_default_license_details
-from src.license.view_tool import ConfirmPostView
-from src.license.utils import build_settings_embed, safe_delete_original_response
+from src.license.view.view_tool import ConfirmPostView
+from src.license.utils import build_settings_embed, safe_delete_original_response, safe_defer, do_simple_owner_id_interaction_check
 
 
 class SettingsView(ui.View):
@@ -19,7 +19,7 @@ class SettingsView(ui.View):
 
     设计模式：
     - 状态自更新：每个开关按钮被点击后，会更新后台数据，然后调用 `update_button_labels` 和
-      `interaction.response.edit_message(view=self)` 来刷新自身，从而在界面上即时反映出
+      `interaction.edit_original_response(view=self)` 来刷新自身，从而在界面上即时反映出
       新的状态（如 ✅ 和 ❌ 的切换），提供了良好的交互反馈。
     - 独立确认流程：对于危险操作（重置、删除数据），它不会直接执行，而是会弹出另一个
       临时的、独立的确认视图（`ConfirmPostView`），防止用户误操作。
@@ -35,11 +35,8 @@ class SettingsView(ui.View):
         self.update_button_labels()  # 初始化时设置正确的按钮标签
 
     async def interaction_check(self, interaction: discord.Interaction) -> bool:
-        # self.config.user_id 是这个设置面板的真正主人
-        if interaction.user.id != self.config.user_id:
-            await interaction.response.send_message("❌ 你不能修改别人的设置。", ephemeral=True)
-            return False
-        return True
+        """权限检查"""
+        return await do_simple_owner_id_interaction_check(self.config.user_id, interaction)
 
     def update_button_labels(self):
         """根据当前的 `self.config` 状态，更新按钮上的标签和表情符号。"""
@@ -57,44 +54,47 @@ class SettingsView(ui.View):
         new_embed = build_settings_embed(self.config)
 
         # 编辑原始消息，同时更新Embed和View
-        await interaction.response.edit_message(embed=new_embed, view=self)
+        await interaction.edit_original_response(embed=new_embed, view=self)
 
     @ui.button(label="切换机器人总开关", style=discord.ButtonStyle.primary, row=0)
     async def toggle_bot_enabled_button(self, interaction: discord.Interaction, button: ui.Button):
         """切换“机器人总开关”选项。"""
+        await safe_defer(interaction)
         self.config.bot_enabled = not self.config.bot_enabled
         await self._update_view(interaction)
 
     @ui.button(label="切换自动发布", style=discord.ButtonStyle.primary, row=0)
     async def toggle_auto_post_button(self, interaction: discord.Interaction, button: ui.Button):
         """切换“自动发布”选项。"""
+        await safe_defer(interaction)
         self.config.auto_post = not self.config.auto_post
         await self._update_view(interaction)
 
     @ui.button(label="切换发布前二次确认", style=discord.ButtonStyle.primary, row=1)
     async def toggle_confirmation_button(self, interaction: discord.Interaction, button: ui.Button):
         """切换“发布前二次确认”选项。"""
+        await safe_defer(interaction)
         self.config.require_confirmation = not self.config.require_confirmation
         await self._update_view(interaction)
 
     @ui.button(label="重置我的协议", style=discord.ButtonStyle.danger, row=2)
     async def reset_license(self, interaction: discord.Interaction, button: ui.Button):
         """重置用户的授权协议为默认值，这是一个危险操作，需要二次确认。"""
-
+        await safe_defer(interaction)
         async def on_confirm(confirm_interaction: discord.Interaction):
             # 确认后，执行重置操作
             self.config.license_details = get_default_license_details(self.config.user_id)
             self.db.save_config(self.config)
-            await confirm_interaction.response.edit_message(content="✅ 你的授权协议已重置为默认值。", embed=None, view=None)
+            await confirm_interaction.edit_original_response(content="✅ 你的授权协议已重置为默认值。", embed=None, view=None)
             await safe_delete_original_response(confirm_interaction, sleep_time=1)
 
         async def on_cancel(cancel_interaction: discord.Interaction):
-            await cancel_interaction.response.edit_message(content="🚫 操作已取消。", embed=None, view=None)
+            await cancel_interaction.edit_original_response(content="🚫 操作已取消。", embed=None, view=None)
             await safe_delete_original_response(cancel_interaction, sleep_time=1)
 
         # 发起一个独立的、临时的确认流程
         confirm_view = ConfirmPostView(interaction.user.id, on_confirm, on_cancel)
-        await interaction.response.send_message(
+        await interaction.followup.send(
             "**⚠️ 警告：** 此操作会将你的默认协议恢复为初始设置！\n请确认你的操作：",
             view=confirm_view,
             ephemeral=True
@@ -103,21 +103,21 @@ class SettingsView(ui.View):
     @ui.button(label="删除所有数据", style=discord.ButtonStyle.danger, row=2)
     async def delete_data(self, interaction: discord.Interaction, button: ui.Button):
         """删除用户在本机器人中的所有数据，这是一个非常危险的操作，需要二次确认。"""
-
+        await safe_defer(interaction)
         async def on_confirm(confirm_interaction: discord.Interaction):
             # 执行真正的删除操作
             try:
                 self.db.delete_config(self.config.user_id)
             except OSError as e:
                 if self.cog.logger: self.cog.logger.error(f"删除用户数据文件失败: {self.config.user_id}, 错误: {e}")
-                await confirm_interaction.response.edit_message(content=f"❌ 删除数据时发生错误！请联系管理员。", view=None)
+                await confirm_interaction.edit_original_response(content=f"❌ 删除数据时发生错误！请联系管理员。", view=None)
                 return
 
             # 在确认面板上给出最终反馈
             cmd_name = ACTIVE_COMMAND_CONFIG["group"]["name"]
             cmd_name_panel = ACTIVE_COMMAND_CONFIG["panel"]["name"]
 
-            await confirm_interaction.response.edit_message(
+            await confirm_interaction.edit_original_response(
                 content=
                 "🗑️ **你的所有数据已被永久删除。**\n"
                 "> **所有的控制面板即将/已经关闭。**\n"
@@ -132,12 +132,12 @@ class SettingsView(ui.View):
                 await self.cog.cleanup_previous_helpers(thread=self.thread)
 
         async def on_cancel(cancel_interaction: discord.Interaction):
-            await cancel_interaction.response.edit_message(content="🚫 操作已取消，你的数据安然无恙。", view=None)
+            await cancel_interaction.edit_original_response(content="🚫 操作已取消，你的数据安然无恙。", view=None)
             await safe_delete_original_response(cancel_interaction, sleep_time=1)
 
         # 创建并发送带有强烈警告的独立确认视图
         confirm_view = ConfirmPostView(interaction.user.id, on_confirm, on_cancel)
-        await interaction.response.send_message(
+        await interaction.followup.send(
             "**🚨 终极警告：此操作不可逆！🚨**\n\n"
             "你确定要**永久删除**你保存在本机器人中的所有数据吗？这包括：\n"
             "- 你保存的默认授权协议\n"
@@ -150,6 +150,6 @@ class SettingsView(ui.View):
     @ui.button(label="关闭面板", style=discord.ButtonStyle.secondary, row=3)
     async def close_panel(self, interaction: discord.Interaction, button: ui.Button):
         """关闭（即删除）这个设置面板消息。"""
-        await interaction.response.defer()  # 先响应，防止超时
+        await safe_defer(interaction)
         await interaction.delete_original_response()
         self.stop()
